Translator_Application/requirements.txt
streamlit
langchain
python-dotenv
google-generativeai